requirements.txt
# App
aiogram
Babel
environs
loguru
apscheduler

# REST API
fastapi
uvicorn
gunicorn
jinja2
pyjwt

# Database & storage
motor
tornado
pydantic
redis
